RecipesPipeline.jl

Documentation for RecipesPipeline.jl

RecipesPipeline.recipe_pipeline! — Method

recipe_pipeline!(plt, plotattributes, args)

Recursively apply user recipes, type recipes, plot recipes and series recipes to build a list of Dicts, each corresponding to a series. At the beginning plotattributes contains only the keyword arguments passed in by the user. Add all series to the plot bject plt and return it.