Recipes processing

RecipesPipeline._process_userrecipes! — Method

_process_userrecipes(plt, plotattributes, args)

Wrap input arguments in a RecipeData' vector and recursively apply user recipes and type recipes on the first element. Prepend the returnedRecipeDatavector. If an element with empyargsis returned pop it from the vector, finish up, and it to vector ofDicts with processed series. When all arguments are processed return the seriesDict`.

RecipesPipeline._process_plotrecipes! — Method

_process_plotrecipes!(plt, kw_list)

Grab the first in line to be processed and pass it through apply_recipe to generate a list of RecipeData objects. If we applied a "plot recipe" without error, then add the returned datalist's KWs, otherwise we just add the original KW.

RecipesPipeline._process_seriesrecipes! — Method

_process_seriesrecipes!(plt, kw_list)

Recursively apply series recipes until the backend supports the seriestype