API

RecipesPipeline.get_axis_limits — Method

get_axis_limits(plt, letter)

Get the limits for the axis specified by letter (:x, :y or :z) in plt. If it errors, tryrange from PlotUtils is used.

RecipesPipeline.is_axis_attribute — Method

is_axis_attribute(plt, attr)

Returns true if attr is an axis attribute, i.e. it applies to xattr, yattr and zattr, otherwise false.

RecipesPipeline.plot_setup! — Method

plot_setup!(plt, plotattributes, kw_list)

Setup plot, subplots and layouts. For example, Plots creates the backend figure, initializes subplots, expands extrema and links subplot axes.

RecipesPipeline.process_sliced_series_attributes! — Method

process_sliced_series_attributes!(plt, kw_list)

All series attributes are now properly resolved. Any change of the kw_list before the application of recipes must come here.

RecipesPipeline.process_userrecipe! — Method

process_userrecipe!(plt, attributes_list, attributes)

Do plotting package specific post-processing and add series attributes to attributeslist. For example, Plots increases the number of series in plt, sets `:seriesplotindex` in attributes and possible adds new series attributes for errorbars or smooth.

RecipesPipeline.slice_series_attributes! — Method

slice_series_attributes!(plt, kw_list, kw)

For attributes given as vector with one element per series, only select the value for current series.

RecipesPipeline.split_attribute — Method

split_attribute(plt, key, val, indices)

Select the proper indices from val for attribute key.

RecipesPipeline.splittable_attribute — Method

splittable_attribute(plt, key, val, len)

Returns true if the attribute key with the value val can be split into groups with group provided as a vector of length len, false otherwise.

RecipesPipeline.type_alias — Method

type_alias(plt, st)

Return the seriestype alias for st.

Utility functions

RecipesPipeline.Formatted — Type

Represents data values with formatting that should apply to the tick labels.

RecipesPipeline.Surface — Type

represents a contour or surface mesh

RecipesPipeline.is3d — Method

is3d(::Type{Val{:myseriestype}})

Returns true if myseriestype represents a 3D series, false otherwise.

RecipesPipeline.is_surface — Method

is_surface(::Type{Val{:myseriestype}})

Returns true if myseriestype represents a surface series, false otherwise.

RecipesPipeline.needs_3d_axes — Method

needs_3d_axes(::Type{Val{:myseriestype}})

Returns true if myseriestype needs 3d axes, false otherwise.